// types/blog.ts
export interface BlogPost {
    title: string;
    date: string;
    slug: string;
    description: string;
    tags: string[];
    category: string;
    coverImage: string;
    author: string;
    readTime: string;
    content: string;
}

export interface BlogPostMeta {
    title: string;
    date: string;
    slug: string;
    description: string;
    tags: string[];
    category: string;
    coverImage: string;
    author: string;
    readTime: string;
}